requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
hyperliquid-python-sdk>=0.1.0
python-dateutil>=2.8.0
requests>=2.28.0

# Optional: For signature verification (uncomment if needed)
# web3>=6.0.0
# eth-account>=0.8.0